feat: refactor OAuth state handling and introduce step-up authentication methods
