+++ Sources/FronteggSwift/FronteggAuth.swift
@@ -41,33 +41,38 @@ public class FronteggAuth: FronteggState {
         return FronteggApp.shared.auth
     }
     
-    private let logger = getLogger("FronteggAuth")
+    let logger = getLogger("FronteggAuth")
     public let credentialManager: CredentialManager
-    private var multiFactorAuthenticator: MultiFactorAuthenticator
-    private var stepUpAuthenticator: StepUpAuthenticator
+    // internal for extension access (FronteggAuth+AdvancedAuth.swift)
+    var multiFactorAuthenticator: MultiFactorAuthenticator
+    var stepUpAuthenticator: StepUpAuthenticator
     public var api: Api
     public var featureFlags: FeatureFlags
     public var entitlements: Entitlements
     private var subscribers = Set<AnyCancellable>()
-    private var refreshTokenDispatch: DispatchWorkItem?
-    private var offlineDebounceWork: DispatchWorkItem?
+    // internal for extension access (Refresh, Testing, Connectivity)
+    var refreshTokenDispatch: DispatchWorkItem?
+    var offlineDebounceWork: DispatchWorkItem?
     private let offlineDebounceDelay: TimeInterval = 0.6
     private let unauthenticatedStartupOfflineCommitWindow: TimeInterval = 4.5
     private let unauthenticatedStartupProbeDelay: TimeInterval = 0.5
     private let unauthenticatedStartupProbeTimeout: TimeInterval = 1.0
     var loginCompletion: CompletionHandler? = nil
-    private var networkMonitoringToken: NetworkStatusMonitor.OnChangeToken?
+    // internal for extension access (Connectivity, Testing, SessionRestore)
+    var networkMonitoringToken: NetworkStatusMonitor.OnChangeToken?
     private var pendingOAuthErrorContext: FronteggOAuthErrorContext?
     private var pendingOAuthErrorPresentationWorkItem: DispatchWorkItem?
     private var pendingEmbeddedOAuthErrorFallbackWorkItem: DispatchWorkItem?
     private let oauthErrorPresentationDelay: TimeInterval = 0.35
     private let embeddedOAuthErrorRecoveryFallbackDelay: TimeInterval = 1.25
-    private var isInitializingWithTokens: Bool = false
-    @MainActor private var isLoginInProgress: Bool = false
-    private let entitlementsLoadLock = NSLock()
-    private var entitlementsLoadInProgress: Bool = false
-    private var entitlementsLoadPendingCompletions: [((Bool) -> Void)] = []
-    private var entitlementsLoadForceRefreshPending: Bool = false
+    // internal for extension access (SessionRestore, Testing, Refresh, HostedFlows)
+    var isInitializingWithTokens: Bool = false
+    @MainActor var isLoginInProgress: Bool = false
+    // internal for extension access (FronteggAuth+Entitlements.swift)
+    let entitlementsLoadLock = NSLock()
+    var entitlementsLoadInProgress: Bool = false
+    var entitlementsLoadPendingCompletions: [((Bool) -> Void)] = []
+    var entitlementsLoadForceRefreshPending: Bool = false
 
     internal static func isUserCancelledOAuthFlow(_ error: Error) -> Bool {
         let nsError = error as NSError
@@ -145,177 +150,8 @@ public class FronteggAuth: FronteggState {
     }
     
     
-    public func manualInit(baseUrl:String, clientId:String, applicationId: String?, entitlementsEnabled: Bool = false) {
-        setLateInit(false)
-        self.baseUrl = baseUrl
-        self.clientId = clientId
-        self.applicationId = applicationId
-        self.isRegional = false
-        self.api = Api(baseUrl: self.baseUrl, clientId: self.clientId, applicationId: self.applicationId)
-        self.featureFlags = FeatureFlags(.init(clientId: self.clientId, api: self.api))
-        self.entitlements = Entitlements(.init(api: self.api, enabled: entitlementsEnabled))
-        resetEntitlementsLoadState()
-        self.initializeSubscriptions()
-    }
-    
-    public func manualInitRegions(regions:[RegionConfig], entitlementsEnabled: Bool = false) {
-        setLateInit(false)
-        self.isRegional = true
-        self.regionData = regions
-        setSelectedRegion(self.getSelectedRegion())
-
-        if let config = self.selectedRegion {
-            self.baseUrl = config.baseUrl
-            self.clientId = config.clientId
-            self.applicationId = config.applicationId
-            self.api = Api(baseUrl: self.baseUrl, clientId: self.clientId, applicationId: self.applicationId)
-            self.featureFlags = FeatureFlags(.init(clientId: self.clientId, api: self.api))
-            self.entitlements = Entitlements(.init(api: self.api, enabled: entitlementsEnabled))
-            resetEntitlementsLoadState()
-            self.initializeSubscriptions()
-        } else {
-            // selectedRegion is nil (e.g. no saved region or invalid selection) – use first region as
-            // fallback so api/credentials are valid. When regions is empty, skip reinit and subscriptions
-            // to avoid using stale api/featureFlags/entitlements.
-            if let fallback = regions.first {
-                self.baseUrl = fallback.baseUrl
-                self.clientId = fallback.clientId
-                self.applicationId = fallback.applicationId
-                self.api = Api(baseUrl: self.baseUrl, clientId: self.clientId, applicationId: self.applicationId)
-                self.featureFlags = FeatureFlags(.init(clientId: self.clientId, api: self.api))
-                self.entitlements = Entitlements(.init(api: self.api, enabled: entitlementsEnabled))
-                resetEntitlementsLoadState()
-                self.initializeSubscriptions()
-            }
-        }
-    }
-
-    /// Resets entitlements load state when entitlements instance is replaced (e.g. manualInit, region switch).
-    /// Invokes any pending completions with false to indicate the load was aborted.
-    private func resetEntitlementsLoadState() {
-        let pending = entitlementsLoadLock.withLock { () -> [((Bool) -> Void)] in
-            entitlementsLoadInProgress = false
-            entitlementsLoadForceRefreshPending = false
-            let pending = entitlementsLoadPendingCompletions
-            entitlementsLoadPendingCompletions.removeAll()
-            return pending
-        }
-        pending.forEach { c in
-            if Thread.isMainThread {
-                c(false)
-            } else {
-                DispatchQueue.main.async { c(false) }
-            }
-        }
-    }
-
-    private func resolveAccessTokenForCurrentUser() -> String? {
-        if let token = self.accessToken, !token.isEmpty { return token }
-        let config = try? PlistHelper.fronteggConfig()
-        let enableSessionPerTenant = config?.enableSessionPerTenant ?? false
-        if enableSessionPerTenant, let tenantId = self.user?.activeTenant.id {
-            return try? credentialManager.getTokenForTenant(tenantId: tenantId, tokenType: .accessToken)
-        }
-        return try? credentialManager.get(key: KeychainKeys.accessToken.rawValue)
-    }
-
-    public func loadEntitlements(forceRefresh: Bool = false, completion: ((Bool) -> Void)? = nil) {
-        func wrapCompletion(_ c: @escaping (Bool) -> Void) -> (Bool) -> Void {
-            return { success in
-                if Thread.isMainThread {
-                    c(success)
-                } else {
-                    DispatchQueue.main.async { c(success) }
-                }
-            }
-        }
-        entitlementsLoadLock.lock()
-        if !forceRefresh && entitlements.hasLoaded {
-            entitlementsLoadLock.unlock()
-            if let c = completion {
-                wrapCompletion(c)(true)
-            }
-            return
-        }
-        if entitlementsLoadInProgress {
-            if forceRefresh {
-                entitlementsLoadForceRefreshPending = true
-            }
-            if let c = completion {
-                entitlementsLoadPendingCompletions.append(wrapCompletion(c))
-            }
-            entitlementsLoadLock.unlock()
-            return
-        }
-        entitlementsLoadInProgress = true
-        if let c = completion {
-            entitlementsLoadPendingCompletions.append(wrapCompletion(c))
-        }
-        entitlementsLoadLock.unlock()
-        performEntitlementsLoad()
-    }
-
-    private enum EntitlementsLoadNextStep {
-        case reload
-        case invoke([((Bool) -> Void)])
-    }
-
-    private func finishEntitlementsLoadCycle() -> EntitlementsLoadNextStep {
-        entitlementsLoadLock.withLock {
-            entitlementsLoadInProgress = false
-            if entitlementsLoadForceRefreshPending {
-                entitlementsLoadForceRefreshPending = false
-                entitlementsLoadInProgress = true
-                return .reload
-            }
-
-            let completions = entitlementsLoadPendingCompletions
-            entitlementsLoadPendingCompletions.removeAll()
-            return .invoke(completions)
-        }
-    }
-
-    private func performEntitlementsLoad() {
-        Task {
-            let success: Bool
-            if let token = resolveAccessTokenForCurrentUser() {
-                success = await entitlements.load(accessToken: token)
-            } else {
-                logger.warning("loadEntitlements: no access token available")
-                success = false
-            }
-
-            switch finishEntitlementsLoadCycle() {
-            case .reload:
-                performEntitlementsLoad()
-            case .invoke(let completions):
-                completions.forEach { $0(success) }
-            }
-        }
-    }
-
-    public func getFeatureEntitlements(featureKey: String) -> Entitlement {
-        guard self.user != nil else {
-            return Entitlement(isEntitled: false, justification: "NOT_AUTHENTICATED")
-        }
-        return entitlements.checkFeature(featureKey: featureKey)
-    }
-
-    public func getPermissionEntitlements(permissionKey: String) -> Entitlement {
-        guard self.user != nil else {
-            return Entitlement(isEntitled: false, justification: "NOT_AUTHENTICATED")
-        }
-        return entitlements.checkPermission(permissionKey: permissionKey)
-    }
-
-    public func getEntitlements(options: EntitledToOptions) -> Entitlement {
-        switch options {
-        case .featureKey(let key):
-            return getFeatureEntitlements(featureKey: key)
-        case .permissionKey(let key):
-            return getPermissionEntitlements(permissionKey: key)
-        }
-    }
+    // MARK: Region Management — see FronteggAuth+RegionManagement.swift
+    // MARK: Entitlements — see FronteggAuth+Entitlements.swift
 
     @MainActor
     @objc private func applicationDidBecomeActive() {
@@ -333,38 +169,6 @@ public class FronteggAuth: FronteggState {
         logger.info("application enter background")
     }
 
-    public func reinitWithRegion(config:RegionConfig) {
-        self.baseUrl = config.baseUrl
-        self.clientId = config.clientId
-        self.applicationId = config.applicationId
-        setSelectedRegion(config)
-        self.api = Api(baseUrl: self.baseUrl, clientId: self.clientId, applicationId: self.applicationId)
-        self.featureFlags = FeatureFlags(.init(clientId: self.clientId, api: self.api))
-        self.entitlements = Entitlements(.init(api: self.api, enabled: FronteggApp.shared.entitlementsEnabled))
-        resetEntitlementsLoadState()
-        loadEntitlements(forceRefresh: true)
-        self.initializeSubscriptions()
-    }
-    
-    
-    public func getSelectedRegion() -> RegionConfig? {
-        guard let selectedRegionKey = CredentialManager.getSelectedRegion() else {
-            return nil
-        }
-        
-        guard let config = self.regionData.first(where: { config in
-            config.key == selectedRegionKey
-        }) else {
-            let keys: String = self.regionData.map { config in
-                config.key
-            }.joined(separator: ", ")
-            logger.critical("invalid region key \(selectedRegionKey). available regions: \(keys)")
-            return nil
-        }
-        
-        return config
-        
-    }
     
     private func warmingWebViewAsync() {
         DispatchQueue.main.async {
@@ -1295,423 +1099,9 @@ public class FronteggAuth: FronteggState {
         refreshTokenDispatch = nil
     }
 
-#if DEBUG
-    @MainActor
-    func resetForTesting(baseUrlOverride: String? = nil) async {
-        cancelScheduledTokenRefresh()
-        offlineDebounceWork?.cancel()
-        offlineDebounceWork = nil
+    // MARK: Testing — see FronteggAuth+Testing.swift
 
-        NetworkStatusMonitor.stopBackgroundMonitoring()
-        if let token = networkMonitoringToken {
-            NetworkStatusMonitor.removeOnChange(token)
-            networkMonitoringToken = nil
-        }
-
-        if let activeSession = WebAuthenticator.shared.session {
-            activeSession.cancel()
-            WebAuthenticator.shared.session = nil
-        }
-        WebAuthenticator.shared.window = nil
-
-        isInitializingWithTokens = false
-        isLoginInProgress = false
-        pendingAppLink = nil
-        activeEmbeddedOAuthFlow = .login
-        loginHint = nil
-        loginCompletion = nil
-        lastAttemptReason = nil
-        webview = nil
-        Self.testNetworkPathAvailabilityOverride = nil
-
-        credentialManager.deleteLastActiveTenantId()
-        credentialManager.clear()
-        CredentialManager.clearPendingOAuthFlows()
-        UserDefaults.standard.removeObject(forKey: KeychainKeys.region.rawValue)
-
-        setIsAuthenticated(false)
-        setUser(nil)
-        setAccessToken(nil)
-        setRefreshToken(nil)
-        setInitializing(false)
-        setShowLoader(false)
-        setAppLink(false)
-        setWebLoading(false)
-        setLoginBoxLoading(false)
-        setIsLoading(false)
-        setIsOfflineMode(false)
-        setRefreshingToken(false)
-        setIsStepUpAuthorization(false)
-        entitlements.clear()
-        resetEntitlementsLoadState()
-
-        await clearWebsiteDataForTesting(baseUrlOverride: baseUrlOverride)
-        URLCache.shared.removeAllCachedResponses()
-    }
-
-    func setTestNetworkPathAvailabilityOverride(_ available: Bool?) {
-        Self.testNetworkPathAvailabilityOverride = available
-    }
-
-    func hasScheduledTokenRefreshForTesting() -> Bool {
-        refreshTokenDispatch != nil
-    }
-
-    @MainActor
-    func isLoginInProgressForTesting() -> Bool {
-        isLoginInProgress
-    }
-#endif
-
-    @MainActor
-    private func clearWebsiteDataForTesting(baseUrlOverride: String?) async {
-        let store = WKWebsiteDataStore.default()
-        let allWebsiteDataTypes = WKWebsiteDataStore.allWebsiteDataTypes()
-
-        guard let host = Self.resolvedTestingResetHost(
-            baseUrlOverride: baseUrlOverride,
-            currentBaseUrl: baseUrl,
-            appBaseUrl: FronteggApp.shared.baseUrl
-        ) else {
-            logger.warning("resetForTesting could not resolve a scoped host, falling back to global WebKit data cleanup")
-            await removeAllWebsiteData(store: store, websiteDataTypes: allWebsiteDataTypes)
-            clearSharedCookiesForTesting(host: nil)
-            return
-        }
-
-        let websiteDataTypes = Self.testingResetWebsiteDataTypes()
-        logger.info("resetForTesting clearing scoped web data for host: \(host)")
-
-        let records = await fetchWebsiteDataRecords(store: store, websiteDataTypes: websiteDataTypes)
-        let matchingRecords = records.filter {
-            Self.shouldRemoveTestingWebsiteDataRecord(named: $0.displayName, forHost: host)
-        }
-
-        if !matchingRecords.isEmpty {
-            await removeWebsiteData(store: store, websiteDataTypes: websiteDataTypes, records: matchingRecords)
-        }
-
-        let storeCookies = await getAllCookies(store: store)
-        for cookie in storeCookies where cookieDomain(cookie.domain, matches: host) {
-            await deleteCookie(store: store, cookie: cookie)
-        }
-
-        clearSharedCookiesForTesting(host: host)
-    }
-
-    @MainActor
-    private func fetchWebsiteDataRecords(
-        store: WKWebsiteDataStore,
-        websiteDataTypes: Set<String>
-    ) async -> [WKWebsiteDataRecord] {
-        await withCheckedContinuation { (continuation: CheckedContinuation<[WKWebsiteDataRecord], Never>) in
-            store.fetchDataRecords(ofTypes: websiteDataTypes) { records in
-                continuation.resume(returning: records)
-            }
-        }
-    }
-
-    @MainActor
-    private func removeWebsiteData(
-        store: WKWebsiteDataStore,
-        websiteDataTypes: Set<String>,
-        records: [WKWebsiteDataRecord]
-    ) async {
-        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-            store.removeData(ofTypes: websiteDataTypes, for: records) {
-                continuation.resume()
-            }
-        }
-    }
-
-    @MainActor
-    private func removeAllWebsiteData(
-        store: WKWebsiteDataStore,
-        websiteDataTypes: Set<String>
-    ) async {
-        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-            store.removeData(ofTypes: websiteDataTypes, modifiedSince: .distantPast) {
-                continuation.resume()
-            }
-        }
-    }
-
-    @MainActor
-    private func getAllCookies(store: WKWebsiteDataStore) async -> [HTTPCookie] {
-        await withCheckedContinuation { (continuation: CheckedContinuation<[HTTPCookie], Never>) in
-            store.httpCookieStore.getAllCookies { cookies in
-                continuation.resume(returning: cookies)
-            }
-        }
-    }
-
-    @MainActor
-    private func deleteCookie(store: WKWebsiteDataStore, cookie: HTTPCookie) async {
-        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-            store.httpCookieStore.delete(cookie) {
-                continuation.resume()
-            }
-        }
-    }
-
-    private func clearSharedCookiesForTesting(host: String?) {
-        let cookiesToDelete: [HTTPCookie]
-        if let host, !host.isEmpty {
-            cookiesToDelete = (HTTPCookieStorage.shared.cookies ?? []).filter {
-                cookieDomain($0.domain, matches: host)
-            }
-        } else {
-            cookiesToDelete = HTTPCookieStorage.shared.cookies ?? []
-        }
-
-        for cookie in cookiesToDelete {
-            HTTPCookieStorage.shared.deleteCookie(cookie)
-        }
-    }
-
-    internal static func resolvedTestingResetHost(
-        baseUrlOverride: String?,
-        currentBaseUrl: String,
-        appBaseUrl: String
-    ) -> String? {
-        for candidate in [baseUrlOverride, currentBaseUrl, appBaseUrl] {
-            guard let candidate,
-                  let host = URL(string: candidate)?.host?.lowercased(),
-                  !host.isEmpty,
-                  host != "late-init.invalid" else {
-                continue
-            }
-            return host
-        }
-        return nil
-    }
-
-    internal static func shouldRemoveTestingWebsiteDataRecord(named displayName: String, forHost host: String) -> Bool {
-        let normalizedDisplayName = displayName.lowercased()
-        let normalizedHost = host.lowercased()
-
-        return normalizedDisplayName == normalizedHost ||
-               normalizedDisplayName.hasSuffix("." + normalizedHost) ||
-               normalizedHost.hasSuffix("." + normalizedDisplayName)
-    }
-
-    internal static func testingResetWebsiteDataTypes() -> Set<String> {
-        var dataTypes: Set<String> = [
-            WKWebsiteDataTypeCookies,
-            WKWebsiteDataTypeDiskCache,
-            WKWebsiteDataTypeMemoryCache,
-            WKWebsiteDataTypeSessionStorage,
-            WKWebsiteDataTypeLocalStorage,
-            WKWebsiteDataTypeIndexedDBDatabases,
-            WKWebsiteDataTypeWebSQLDatabases,
-            WKWebsiteDataTypeOfflineWebApplicationCache,
-        ]
-
-        if #available(iOS 11.3, *) {
-            dataTypes.insert(WKWebsiteDataTypeFetchCache)
-            dataTypes.insert(WKWebsiteDataTypeServiceWorkerRegistrations)
-        }
-
-        if #available(iOS 16.0, *) {
-            dataTypes.insert(WKWebsiteDataTypeFileSystem)
-        }
-
-        return dataTypes
-    }
-
-    public func logout(clearCookie: Bool = true, _ completion: FronteggAuth.LogoutHandler? = nil) {
-        Task { @MainActor in
-            
-            setIsLoading(true)
-            
-            // Try to reload from keychain if in-memory token is nil
-            // This ensures we can invalidate the session on the server even if the token
-            // was not loaded into memory (e.g., after app restart)
-            let accessTokenForServerLogout = self.accessToken
-            var refreshTokenForServerLogout = self.refreshToken
-            if refreshTokenForServerLogout == nil {
-                if let keychainToken = try? credentialManager.get(key: KeychainKeys.refreshToken.rawValue) {
-                    self.logger.info("Reloaded refresh token from keychain for logout")
-                    refreshTokenForServerLogout = keychainToken
-                } else {
-                    self.logger.warning("No refresh token found in memory or keychain. Server session may remain active.")
-                }
-            }
-
-            // Preserve lastActiveTenantId when enableSessionPerTenant is enabled
-            // This ensures each device maintains its own tenant context even after logout
-            let config = try? PlistHelper.fronteggConfig()
-            let enableSessionPerTenant = config?.enableSessionPerTenant ?? false
-            
-            if enableSessionPerTenant {
-                let preservedTenantId = credentialManager.getLastActiveTenantId()
-                self.logger.info("🔵 [SessionPerTenant] Preserving lastActiveTenantId (\(preservedTenantId ?? "nil")) for per-tenant session isolation")
-                // Clear all items except lastActiveTenantId
-                self.credentialManager.clear(excludingKeys: [KeychainKeys.lastActiveTenantId.rawValue])
-                self.logger.info("🔵 [SessionPerTenant] Cleared keychain while preserving lastActiveTenantId")
-            } else {
-                self.credentialManager.deleteLastActiveTenantId()
-                self.credentialManager.clear()
-            }
-            CredentialManager.clearPendingOAuthFlows()
-
-            if clearCookie {
-                await self.clearCookie()
-            }
-            
-            setIsAuthenticated(false)
-            setUser(nil)
-            setAccessToken(nil)
-            setRefreshToken(nil)
-            setInitializing(false)
-            setAppLink(false)
-            setIsOfflineMode(false)
-            self.isLoginInProgress = false
-            setRefreshingToken(false)
-            setIsStepUpAuthorization(false)
-            self.lastAttemptReason = nil
-            entitlements.clear()
-
-            // Cancel scheduled tasks and stop monitoring
-            cancelScheduledTokenRefresh()
-            offlineDebounceWork?.cancel()
-            offlineDebounceWork = nil
-            NetworkStatusMonitor.stopBackgroundMonitoring()
-            if let token = self.networkMonitoringToken {
-                NetworkStatusMonitor.removeOnChange(token)
-                self.networkMonitoringToken = nil
-            }
-
-            setIsLoading(false)
-            completion?(.success(true))
-
-            Task {
-                await self.api.logout(accessToken: accessTokenForServerLogout, refreshToken: refreshTokenForServerLogout)
-            }
-        }
-
-    }
-    
-    /// Returns true if `domain` matches `host`, supporting leading dot and subdomains.
-    @inline(__always)
-    func cookieDomain(_ domain: String, matches host: String) -> Bool {
-        let cd = domain.hasPrefix(".") ? String(domain.dropFirst()) : domain
-        return host == cd || host.hasSuffix("." + cd)
-    }
-    /// Builds a regex-based name matcher using `self.cookieRegex`.
-    /// Falls back to `^fe_refresh` if empty or invalid.
-    func makeCookieNameMatcher() -> (String) -> Bool {
-        let fallback = "^fe_refresh"
-        
-        var cookieRegex: String?
-        if let config = try? PlistHelper.fronteggConfig() {
-            cookieRegex = config.cookieRegex
-        }
-        
-        let pattern = (cookieRegex != nil && cookieRegex?.isEmpty == false) ? cookieRegex! : fallback
-        do {
-            let re = try NSRegularExpression(pattern: pattern)
-            return { name in
-                let range = NSRange(name.startIndex..<name.endIndex, in: name)
-                return re.firstMatch(in: name, range: range) != nil
-            }
-        } catch {
-            self.logger.warning("Invalid cookie regex '\(pattern)'. Using fallback '\(fallback)'. Error: \(error.localizedDescription)")
-            let re = try! NSRegularExpression(pattern: fallback)
-            return { name in
-                let range = NSRange(name.startIndex..<name.endIndex, in: name)
-                return re.firstMatch(in: name, range: range) != nil
-            }
-        }
-    }
-    
-    
-    /// Deletes cookies that match the configured name regex and (optionally) the current host.
-    /// - Behavior:
-    ///   - If `deleteCookieForHostOnly == true`, restricts deletion to cookies whose domain matches `baseUrl`'s host.
-    ///   - If `deleteCookieForHostOnly == false`, deletes any cookie whose name matches the regex (domain-agnostic).
-    /// - Awaited to guarantee that deletion completes before continuing logout flow.
-    @MainActor
-    func clearCookie() async {
-        
-        var deleteCookieForHostOnly: Bool = true
-        var cookieRegex: String?
-        if let config = try? PlistHelper.fronteggConfig() {
-            deleteCookieForHostOnly = config.deleteCookieForHostOnly
-            cookieRegex = config.cookieRegex
-        }
-        
-        let restrictToHost = deleteCookieForHostOnly
-        
-        // Resolve host only when needed
-        let host: String? = {
-            guard restrictToHost else { return nil }
-            guard let h = URL(string: baseUrl)?.host else {
-                logger.warning("Invalid baseUrl; cannot resolve host. Proceeding without domain restriction.")
-                return nil
-            }
-            return h
-        }()
-        
-        let store = WKWebsiteDataStore.default().httpCookieStore
-        
-        // Fetch all cookies
-        let cookies: [HTTPCookie] = await withCheckedContinuation { (cont: CheckedContinuation<[HTTPCookie], Never>) in
-            store.getAllCookies { cont.resume(returning: $0) }
-        }
-        
-        // Deduplicate defensively (name+domain+path is the natural identity)
-        let uniqueCookies: [HTTPCookie] = {
-            var seen = Set<String>()
-            return cookies.filter { c in
-                let key = "\(c.name)|\(c.domain)|\(c.path)"
-                return seen.insert(key).inserted
-            }
-        }()
-        
-        let nameMatches = makeCookieNameMatcher()
-        
-        // Compose predicate
-        let shouldDelete: (HTTPCookie) -> Bool = { cookie in
-            guard nameMatches(cookie.name) else { return false }
-            guard let h = host else { return true } // no domain restriction
-            let match = self.cookieDomain(cookie.domain, matches: h)
-            if !match {
-                self.logger.debug("Skipping cookie due to domain mismatch: \(cookie.name) @ \(cookie.domain) (host: \(h))")
-            }
-            return match
-        }
-        
-        let targets = uniqueCookies.filter(shouldDelete)
-        
-        guard targets.isEmpty == false else {
-            self.logger.debug("No cookies matched for deletion. regex: \(cookieRegex ?? "^fe_refresh"), restrictToHost: \(restrictToHost), host: \(host ?? "n/a")")
-            return
-        }
-        
-        // Delete sequentially (deterministic, avoids overloading store). If you prefer parallel, see comment below.
-        var deleted = 0
-        let start = Date()
-        
-        for cookie in targets {
-            await withCheckedContinuation { (cont: CheckedContinuation<Void, Never>) in
-                store.delete(cookie) {
-                    deleted += 1
-                    self.logger.info("Deleted cookie [\(deleted)/\(targets.count)]: \(cookie.name) @ \(cookie.domain)\(cookie.path)")
-                    cont.resume()
-                }
-            }
-        }
-        
-        let elapsed = String(format: "%.2fs", Date().timeIntervalSince(start))
-        self.logger.info("Cookie cleanup completed. Deleted \(deleted)/\(targets.count) cookies in \(elapsed).")
-    }
-    
-    public func logout() {
-        logout { res in
-            self.logger.info("Logged out")
-        }
-    }
+    // MARK: Logout — see FronteggAuth+Logout.swift
     
     private func unwrapURLError(_ error: Error) -> URLError? {
         // Walk underlying errors to find a URLError
@@ -3217,14 +2607,8 @@ public class FronteggAuth: FronteggState {
             }
         }
     }
-    private func pendingOAuthState(from url: URL) -> String? {
-        guard let components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return nil
-        }
-
-        return components.queryItems?.first(where: { item in
-            item.name == "state" && !(item.value?.isEmpty ?? true)
-        })?.value
+    func pendingOAuthState(from url: URL) -> String? {
+        CredentialManager.pendingOAuthState(from: url)
     }
 
     private func matchesGeneratedRedirectUri(_ url: URL) -> Bool {
@@ -4415,264 +3799,6 @@ public class FronteggAuth: FronteggState {
         }
     }
     
-    public func loginWithPasskeys (_ _completion: FronteggAuth.CompletionHandler? = nil){
-        
-        if #available(iOS 15.0, *) {
-            Task {
-                let completion = handleMfaRequired(_completion)
-                await PasskeysAuthenticator.shared.loginWithPasskeys(completion)
-            }
-        } else {
-            // Fallback on earlier versions
-        }
-    }
-    public func registerPasskeys(_ completion: FronteggAuth.ConditionCompletionHandler? = nil) {
-        
-        if #available(iOS 15.0, *) {
-            PasskeysAuthenticator.shared.startWebAuthn(completion)
-        } else {
-            // Fallback on earlier versions
-        }
-    }
-    
-    
-    public func handleSocialLoginCallback(_ url: URL) -> URL? {
-        guard let comps = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return nil
-        }
-        
-        // 1) Host must match Frontegg base URL host
-        guard let allowedHost = URL(string: FronteggAuth.shared.baseUrl)?.host else {
-            return nil
-        }
-        
-        guard comps.host == allowedHost else {
-            return nil
-        }
-        
-        // 2) Path: /oauth/account/redirect/ios/{bundleId}/{provider}
-        let prefix = "/oauth/account/redirect/ios/"
-        let path = comps.path
-        guard path.hasPrefix(prefix) || path.hasPrefix("/ios/oauth/callback") else {
-            return nil
-        }
-        
-        let bundleId = FronteggApp.shared.bundleIdentifier
-        guard !bundleId.isEmpty else {
-            return nil
-        }
-        
-        // Helpers
-        let items = comps.queryItems ?? []
-        func q(_ name: String) -> String? { items.first { $0.name == name }?.value }
+    // MARK: Advanced Auth — see FronteggAuth+AdvancedAuth.swift
 
-        if q("error") != nil || q("error_description") != nil {
-            return nil
-        }
-        
-        // Extract supported params
-        var queryParams: [String: String] = [:]
-        
-        if let code = q("code"), !code.isEmpty {
-            queryParams["code"] = code
-        }
-        
-        if let idToken = q("id_token"), !idToken.isEmpty {
-            queryParams["id_token"] = idToken
-        }
-        
-        let redirectUri = SocialLoginUrlGenerator.shared.defaultRedirectUri()
-        queryParams["redirectUri"] = redirectUri.addingPercentEncoding(withAllowedCharacters: .alphanumerics)
-        
-        // Process state
-        if let state = q("state"), !state.isEmpty {
-            if let data = state.data(using: .utf8),
-               var dict = try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any] {
-                dict.removeValue(forKey: "platform")
-                dict.removeValue(forKey: "bundleId")
-                if let newData = try? JSONSerialization.data(withJSONObject: dict, options: []),
-                   let newState = String(data: newData, encoding: .utf8) {
-                    queryParams["state"] = newState.addingPercentEncoding(withAllowedCharacters: .alphanumerics)
-                }
-            } else {
-                // fallback if state is not valid JSON
-                queryParams["state"] = state
-            }
-        }
-        
-        if let s = WebAuthenticator.shared.session {
-            s.cancel()
-        }
-        
-        // Build query string safely
-        var compsOut = URLComponents()
-        compsOut.queryItems = queryParams.map { URLQueryItem(name: $0.key, value: $0.value) }
-        
-        let finalUrl = URL(string: "\(FronteggAuth.shared.baseUrl)/oauth/account/social/success?\(compsOut.query ?? "")")
-        
-        return finalUrl
-    }
-    
-    
-    
-    /// Authorizes the session using a refresh token (and optional device token cookie).
-    /// The refresh token must come from identity-server APIs, e.g. sign-up:
-    /// `POST /frontegg/identity/resources/users/v1/signUp`.
-    public func requestAuthorizeAsync(refreshToken: String, deviceTokenCookie: String? = nil) async throws -> User {
-        FronteggAuth.shared.setIsLoading(true)
-        
-        self.logger.info("Requesting authorize with refresh and device tokens")
-        
-        do {
-            let authResponse = try await self.api.authroizeWithTokens(refreshToken: refreshToken, deviceTokenCookie: deviceTokenCookie)
-            await FronteggAuth.shared.setCredentials(accessToken: authResponse.access_token, refreshToken: authResponse.refresh_token)
-            
-            if let user = self.user {
-                return user
-            }
-            
-            throw FronteggError.authError(.failedToAuthenticate)
-        } catch {
-            self.logger.error("Authorization request failed: \(error.localizedDescription)")
-            FronteggAuth.shared.setIsLoading(false)
-            throw error
-        }
-    }
-    
-    /// Callback-based variant of `requestAuthorizeAsync`. Use with tokens from identity-server APIs
-    /// (e.g. `POST /frontegg/identity/resources/users/v1/signUp`).
-    public func requestAuthorize(refreshToken: String, deviceTokenCookie: String? = nil, _ completion: @escaping FronteggAuth.CompletionHandler) {
-        DispatchQueue.global(qos: .userInitiated).async {
-            Task {
-                do {
-                    let user = try await self.requestAuthorizeAsync(refreshToken: refreshToken, deviceTokenCookie: deviceTokenCookie)
-                    await MainActor.run {
-                        completion(.success(user)) // Assuming success is represented by empty parentheses
-                    }
-                } catch let error as FronteggError {
-                    await MainActor.run {
-                        completion(.failure(error))
-                    }
-                } catch {
-                    self.logger.error("Failed to authenticate: \(error.localizedDescription)")
-                    await MainActor.run {
-                        completion(.failure(.authError(.failedToAuthenticate)))
-                    }
-                }
-            }
-        }
-    }
-    
-    /// Checks if the user has been stepped up (re-authenticated with stronger authentication).
-    ///
-    /// - Parameter maxAge: Optional parameter to specify the maximum age of the authentication.
-    /// - Returns: `true` if the user has been stepped up, otherwise `false`.
-    public func isSteppedUp(maxAge: TimeInterval? = nil) -> Bool {
-        return self.stepUpAuthenticator.isSteppedUp(maxAge: maxAge)
-    }
-    
-    /// Initiates a step-up authentication process.
-    ///
-    /// This function triggers a step-up authentication process, which requires the user to re-authenticate with stronger authentication mechanisms.
-    ///
-    /// - Parameters:
-    ///   - maxAge: Optional parameter to specify the maximum age of the authentication.
-    ///   - _completion: Optional completion handler that is executed once the step-up process is complete.
-    public func stepUp(
-        maxAge: TimeInterval? = nil,
-        _ _completion: FronteggAuth.CompletionHandler? = nil
-    ) async {
-        return self.stepUpAuthenticator.stepUp(maxAge: maxAge, completion: _completion)
-    }
-    
-    internal func handleMfaRequired(_ _completion: FronteggAuth.CompletionHandler? = nil) -> FronteggAuth.CompletionHandler {
-        let completion: FronteggAuth.CompletionHandler =  { (result) in
-            
-            switch (result) {
-            case .success(_):
-                DispatchQueue.main.async {
-                    FronteggAuth.shared.setIsLoading(false)
-                    _completion?(result)
-                }
-                
-            case .failure(let fronteggError):
-                
-                switch fronteggError {
-                case .authError(let authError):
-                    if case let .mfaRequired(jsonResponse, refreshToken) = authError {
-                        // Handle the MFA-required logic here with the jsonResponse
-                        self.logger.info("MFA required with JSON response: \(jsonResponse)")
-                        self.startMultiFactorAuthenticator(
-                            mfaRequestData: jsonResponse,
-                            refreshToken:refreshToken,
-                            completion: _completion
-                        )
-                        
-                        return
-                    } else {
-                        self.logger.info("authentication error: \(authError.localizedDescription)")
-                    }
-                case .configError(let configError):
-                    self.logger.info("config error: \(configError.localizedDescription)")
-                case .networkError(let error):
-                    self.logger.info("network error: \(error.localizedDescription)")
-                }
-                
-                DispatchQueue.main.async {
-                    FronteggAuth.shared.setIsLoading(false)
-                    _completion?(result)
-                }
-            }
-        }
-        return completion
-    }
-    
-    
-    internal func startMultiFactorAuthenticator(
-        mfaRequestData: [String: Any]? = nil,
-        mfaRequestJson: String? = nil,
-        refreshToken: String? = nil,
-        completion: FronteggAuth.CompletionHandler? = nil
-    ) {
-        Task {
-            do {
-                var authorizeUrl: URL
-
-                if let requestData = mfaRequestData {
-                    (authorizeUrl, _) = try await multiFactorAuthenticator.start(mfaRequestData: requestData, refreshToken: refreshToken)
-                } else if let requestJson = mfaRequestJson {
-                    (authorizeUrl, _) = try multiFactorAuthenticator.start(mfaRequestJson: requestJson)
-                } else {
-                    return
-                }
-                
-                await MainActor.run { [weak self] in
-                    guard let self else { return }
-                    self.setIsLoading(false)
-                    
-                    if self.embeddedMode {
-                        self.activeEmbeddedOAuthFlow = .mfa
-                        self.pendingAppLink = authorizeUrl
-                        self.setWebLoading(true)
-                        self.embeddedLogin(completion, loginHint: nil)
-                        return
-                    }
-                    
-                    let oauthCallback = self.createOauthCallbackHandler(
-                        completion ?? { _ in },
-                        pendingOAuthState: self.pendingOAuthState(from: authorizeUrl),
-                        flow: .mfa
-                    )
-                    WebAuthenticator.shared.start(authorizeUrl, completionHandler: oauthCallback)
-                }
-            } catch let error as FronteggError {
-                await MainActor.run { [weak self] in
-                    guard let self else { return }
-                    self.setIsLoading(false)
-                }
-                completion?(.failure(error))
-            }
-        }
-    }
-    
 }
