+++ Sources/FronteggSwift/auth/FronteggAuth+Logout.swift
@@ -1,0 +1,206 @@
+//
+//  FronteggAuth+Logout.swift
+//
+//  Created by David Frontegg on 14/11/2022.
+//
+
+import Foundation
+import WebKit
+
+extension FronteggAuth {
+
+    public func logout(clearCookie: Bool = true, _ completion: FronteggAuth.LogoutHandler? = nil) {
+        Task { @MainActor in
+
+            setIsLoading(true)
+
+            // Try to reload from keychain if in-memory token is nil
+            // This ensures we can invalidate the session on the server even if the token
+            // was not loaded into memory (e.g., after app restart)
+            let accessTokenForServerLogout = self.accessToken
+            var refreshTokenForServerLogout = self.refreshToken
+            if refreshTokenForServerLogout == nil {
+                if let keychainToken = try? credentialManager.get(key: KeychainKeys.refreshToken.rawValue) {
+                    self.logger.info("Reloaded refresh token from keychain for logout")
+                    refreshTokenForServerLogout = keychainToken
+                } else {
+                    self.logger.warning("No refresh token found in memory or keychain. Server session may remain active.")
+                }
+            }
+
+            // Preserve lastActiveTenantId when enableSessionPerTenant is enabled
+            // This ensures each device maintains its own tenant context even after logout
+            let config = try? PlistHelper.fronteggConfig()
+            let enableSessionPerTenant = config?.enableSessionPerTenant ?? false
+
+            if enableSessionPerTenant {
+                let preservedTenantId = credentialManager.getLastActiveTenantId()
+                self.logger.info("🔵 [SessionPerTenant] Preserving lastActiveTenantId (\(preservedTenantId ?? "nil")) for per-tenant session isolation")
+                // Clear all items except lastActiveTenantId
+                self.credentialManager.clear(excludingKeys: [KeychainKeys.lastActiveTenantId.rawValue])
+                self.logger.info("🔵 [SessionPerTenant] Cleared keychain while preserving lastActiveTenantId")
+            } else {
+                self.credentialManager.deleteLastActiveTenantId()
+                self.credentialManager.clear()
+            }
+            CredentialManager.clearPendingOAuthFlows()
+
+            if clearCookie {
+                await self.clearCookie()
+            }
+
+            setIsAuthenticated(false)
+            setUser(nil)
+            setAccessToken(nil)
+            setRefreshToken(nil)
+            setInitializing(false)
+            setAppLink(false)
+            setIsOfflineMode(false)
+            self.isLoginInProgress = false
+            setRefreshingToken(false)
+            setIsStepUpAuthorization(false)
+            self.lastAttemptReason = nil
+            entitlements.clear()
+
+            // Cancel scheduled tasks and stop monitoring
+            cancelScheduledTokenRefresh()
+            offlineDebounceWork?.cancel()
+            offlineDebounceWork = nil
+            NetworkStatusMonitor.stopBackgroundMonitoring()
+            if let token = self.networkMonitoringToken {
+                NetworkStatusMonitor.removeOnChange(token)
+                self.networkMonitoringToken = nil
+            }
+
+            setIsLoading(false)
+            completion?(.success(true))
+
+            Task {
+                await self.api.logout(accessToken: accessTokenForServerLogout, refreshToken: refreshTokenForServerLogout)
+            }
+        }
+
+    }
+
+    /// Returns true if `domain` matches `host`, supporting leading dot and subdomains.
+    @inline(__always)
+    func cookieDomain(_ domain: String, matches host: String) -> Bool {
+        let cd = domain.hasPrefix(".") ? String(domain.dropFirst()) : domain
+        return host == cd || host.hasSuffix("." + cd)
+    }
+
+    /// Builds a regex-based name matcher using `self.cookieRegex`.
+    /// Falls back to `^fe_refresh` if empty or invalid.
+    func makeCookieNameMatcher() -> (String) -> Bool {
+        let fallback = "^fe_refresh"
+
+        var cookieRegex: String?
+        if let config = try? PlistHelper.fronteggConfig() {
+            cookieRegex = config.cookieRegex
+        }
+
+        let pattern = (cookieRegex != nil && cookieRegex?.isEmpty == false) ? cookieRegex! : fallback
+        do {
+            let re = try NSRegularExpression(pattern: pattern)
+            return { name in
+                let range = NSRange(name.startIndex..<name.endIndex, in: name)
+                return re.firstMatch(in: name, range: range) != nil
+            }
+        } catch {
+            self.logger.warning("Invalid cookie regex '\(pattern)'. Using fallback '\(fallback)'. Error: \(error.localizedDescription)")
+            let re = try! NSRegularExpression(pattern: fallback)
+            return { name in
+                let range = NSRange(name.startIndex..<name.endIndex, in: name)
+                return re.firstMatch(in: name, range: range) != nil
+            }
+        }
+    }
+
+
+    /// Deletes cookies that match the configured name regex and (optionally) the current host.
+    /// - Behavior:
+    ///   - If `deleteCookieForHostOnly == true`, restricts deletion to cookies whose domain matches `baseUrl`'s host.
+    ///   - If `deleteCookieForHostOnly == false`, deletes any cookie whose name matches the regex (domain-agnostic).
+    /// - Awaited to guarantee that deletion completes before continuing logout flow.
+    @MainActor
+    func clearCookie() async {
+
+        var deleteCookieForHostOnly: Bool = true
+        var cookieRegex: String?
+        if let config = try? PlistHelper.fronteggConfig() {
+            deleteCookieForHostOnly = config.deleteCookieForHostOnly
+            cookieRegex = config.cookieRegex
+        }
+
+        let restrictToHost = deleteCookieForHostOnly
+
+        // Resolve host only when needed
+        let host: String? = {
+            guard restrictToHost else { return nil }
+            guard let h = URL(string: baseUrl)?.host else {
+                logger.warning("Invalid baseUrl; cannot resolve host. Proceeding without domain restriction.")
+                return nil
+            }
+            return h
+        }()
+
+        let store = WKWebsiteDataStore.default().httpCookieStore
+
+        // Fetch all cookies
+        let cookies: [HTTPCookie] = await withCheckedContinuation { (cont: CheckedContinuation<[HTTPCookie], Never>) in
+            store.getAllCookies { cont.resume(returning: $0) }
+        }
+
+        // Deduplicate defensively (name+domain+path is the natural identity)
+        let uniqueCookies: [HTTPCookie] = {
+            var seen = Set<String>()
+            return cookies.filter { c in
+                let key = "\(c.name)|\(c.domain)|\(c.path)"
+                return seen.insert(key).inserted
+            }
+        }()
+
+        let nameMatches = makeCookieNameMatcher()
+
+        // Compose predicate
+        let shouldDelete: (HTTPCookie) -> Bool = { cookie in
+            guard nameMatches(cookie.name) else { return false }
+            guard let h = host else { return true } // no domain restriction
+            let match = self.cookieDomain(cookie.domain, matches: h)
+            if !match {
+                self.logger.debug("Skipping cookie due to domain mismatch: \(cookie.name) @ \(cookie.domain) (host: \(h))")
+            }
+            return match
+        }
+
+        let targets = uniqueCookies.filter(shouldDelete)
+
+        guard targets.isEmpty == false else {
+            self.logger.debug("No cookies matched for deletion. regex: \(cookieRegex ?? "^fe_refresh"), restrictToHost: \(restrictToHost), host: \(host ?? "n/a")")
+            return
+        }
+
+        // Delete sequentially (deterministic, avoids overloading store).
+        var deleted = 0
+        let start = Date()
+
+        for cookie in targets {
+            await withCheckedContinuation { (cont: CheckedContinuation<Void, Never>) in
+                store.delete(cookie) {
+                    deleted += 1
+                    self.logger.info("Deleted cookie [\(deleted)/\(targets.count)]: \(cookie.name) @ \(cookie.domain)\(cookie.path)")
+                    cont.resume()
+                }
+            }
+        }
+
+        let elapsed = String(format: "%.2fs", Date().timeIntervalSince(start))
+        self.logger.info("Cookie cleanup completed. Deleted \(deleted)/\(targets.count) cookies in \(elapsed).")
+    }
+
+    public func logout() {
+        logout { res in
+            self.logger.info("Logged out")
+        }
+    }
+}
